+++ HISTORY.md
@@ -66,6 +66,8 @@
 
   * Add Hinge Embedding Loss Function (#2229).
 
+  * Add Margin Ranking Loss Function (#2264).
+
 ### mlpack 3.2.2
 ###### 2019-11-26
   * Add `valid` and `same` padding option in `Convolution` and `Atrous